requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.1.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.35
typing-extensions==4.12.2
werkzeug==3.0.4
zipp==3.20.1
